refactor: move dto for creating a subscription to its own feature directory
