+++ src/walking_skeleton/src/feature/add/AddDTO.ts
@@ -1,0 +1,13 @@
+import type {Price, Rating, RenewalPeriod} from "@/domain/domain.ts";
+
+export interface CreateSubscriptionDTO {
+  name: string;
+  price: Price;
+  /**
+   * Abo-Abschluss, standardisiert
+   */
+  startDate: string,
+  provider: string,
+  renewalPeriod: RenewalPeriod,
+  categories: string[],
+}
