+++ src/walking_skeleton/src/feature/list/dto/ListDTO.ts
@@ -1,16 +1,5 @@
 import type {Price, Rating, RenewalPeriod} from "@/domain/domain.ts";
 
-export interface CreateSubscriptionDTO {
-  name: string;
-  price: Price;
-  /**
-   * Abo-Abschluss, standardisiert
-   */
-  startDate: string,
-  provider: string,
-  renewalPeriod: RenewalPeriod,
-  categories: string[],
-}
 
 export interface SubscriptionListItemDTO {
   id: string,
